Move the getExpression implementation onto the core parser.
